src/panda_nlp/panda_nlp/requirements.txt
openai-whisper==20231117
spacy==3.7.2
sounddevice==0.4.6
numpy==1.24.3
wave
